refactor: Product Repository의 패키지 변경
