+++ src/main/java/cart/repository/product/ProductDbRepository.java
@@ -1,4 +1,4 @@
-package cart.repository;
+package cart.repository.product;
 
 import cart.domain.product.Product;
 import org.springframework.beans.factory.annotation.Autowired;
